src/requirements.txt
fastapi==0.110.0
httpx==0.27.0
starlette==0.36.3
cloudflare-workers==0.0.1
